Add suport to second clustering function

+++ load_grec.m
@@ -16,7 +16,7 @@ function [ graphs_train, clss_train, graphs_valid, clss_valid,...
         graphs_train(i).am = zeros(size(graphs_train(i).v,1),size(graphs_train(i).v,1)) ;
         ind = sub2ind(size(graphs_train(i).am),graphs_train(i).e(:,1),graphs_train(i).e(:,2)) ;
         graphs_train(i).am(ind) = 1 ;
-        graphs_train(i).am = graphs_train(i).am |graphs_train(i).am';
+        graphs_train(i).am = double(graphs_train(i).am |graphs_train(i).am');
     end;
 
     for i = 1:nvalid
@@ -25,7 +25,7 @@ function [ graphs_train, clss_train, graphs_valid, clss_valid,...
         graphs_valid(i).am = zeros(size(graphs_valid(i).v,1),size(graphs_valid(i).v,1)) ;
         ind = sub2ind(size(graphs_valid(i).am),graphs_valid(i).e(:,1),graphs_valid(i).e(:,2)) ;
         graphs_valid(i).am(ind) = 1 ;
-        graphs_valid(i).am = graphs_valid(i).am |graphs_valid(i).am';
+        graphs_valid(i).am = double(graphs_valid(i).am |graphs_valid(i).am');
     end;
 
     for i = 1:ntest
@@ -34,7 +34,7 @@ function [ graphs_train, clss_train, graphs_valid, clss_valid,...
         graphs_test(i).am = zeros(size(graphs_test(i).v,1),size(graphs_test(i).v,1)) ;
         ind = sub2ind(size(graphs_test(i).am),graphs_test(i).e(:,1),graphs_test(i).e(:,2)) ;
         graphs_test(i).am(ind) = 1 ;
-        graphs_test(i).am = graphs_test(i).am |graphs_test(i).am';
+        graphs_test(i).am = double(graphs_test(i).am |graphs_test(i).am');
     end;
 
 end
